backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
anthropic==0.7.7
openai==1.3.7
Pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
